+++ app/overrides/projects/form.rb
@@ -33,8 +33,21 @@ Deface::Override.new(
 )
 Deface::Override.new(
   virtual_path: 'projects/_form',
-  name: 'disable-trackers',
-  replace: "erb[silent]:contains('if @project.new_record? || @project.module_enabled?')",
-  text: "<% if (@project.new_record? || @project.module_enabled?('issue_tracking')) && (Rails.env == 'test') %>",
+  name: 'readonly-trackers',
+  replace: "erb[loud]:contains('@project.trackers.to_a.include?(tracker)')",
+  text: "<%= check_box_tag 'project[tracker_ids][]', tracker.id, @project.trackers.to_a.include?(tracker), :id => nil, :readonly => true %>",
+  #replace: "erb[silent]:contains('if @project.new_record? || @project.module_enabled?')",
+  #text: "<% if (@project.new_record? || @project.module_enabled?('issue_tracking')) && (Rails.env == 'test') %>",
+  namespaced: true
+)
+Deface::Override.new(
+  virtual_path: 'projects/_form',
+  name: 'readonly-custom-fields',
+  replace: "erb[loud]:contains('@project.all_issue_custom_fields.include?')",
+  text: "<%= check_box_tag 'project[issue_custom_field_ids][]', custom_field.id, (@project.all_issue_custom_fields.include? custom_field),
+        :disabled => (custom_field.is_for_all? ? 'disabled' : nil),
+        :id => nil, :readonly => true %>",
+  #replace: "erb[silent]:contains('if @project.new_record? || @project.module_enabled?')",
+  #text: "<% if (@project.new_record? || @project.module_enabled?('issue_tracking')) && (Rails.env == 'test') %>",
   namespaced: true
 )